support multiple inner senders
